Add new geometry buffering/batching
